requirements.txt
streamlit
openai
supabase
Pillow